requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
Pillow==11.0.0
hachoir==3.2.0